Allow passing an ordered array of groups
Some ipp operations such as get-jobs requires serializing more than one
of the same attribute group tag.  The previous model didn't allow
sending multiple as the attribute tag was the key to the object which is
passed to the serializer.  The serializer will now check if the item in
the object is an array and then add each item of the array, preserving
order.